+++ src/processTailwindFeatures.js
@@ -16,8 +16,6 @@ import processPlugins from './util/processPlugins'
 import cloneNodes from './util/cloneNodes'
 import { issueFlagNotices, flagEnabled } from './featureFlags.js'
 
-import darkModeVariantPlugin from './flagged/darkModeVariantPlugin'
-
 import hash from 'object-hash'
 import log from './util/log'
 
@@ -42,11 +40,7 @@ export default function (getConfig) {
       issueFlagNotices(config)
 
       processedPlugins = processPlugins(
-        [
-          ...corePlugins(config),
-          ...[flagEnabled(config, 'darkModeVariant') ? darkModeVariantPlugin : () => {}],
-          ..._.get(config, 'plugins', []),
-        ],
+        [...corePlugins(config), ..._.get(config, 'plugins', [])],
         config
       )
