+++ src/util/getAllConfigs.js
@@ -3,7 +3,6 @@ import { flagEnabled } from '../featureFlags'
 import uniformColorPalette from '../flagged/uniformColorPalette.js'
 import extendedSpacingScale from '../flagged/extendedSpacingScale.js'
 import extendedFontSizeScale from '../flagged/extendedFontSizeScale.js'
-import darkModeVariant from '../flagged/darkModeVariant.js'
 import additionalBreakpoint from '../flagged/additionalBreakpoint'
 import { flatMap, get } from 'lodash'
 
@@ -16,7 +15,6 @@ export default function getAllConfigs(config, defaultPresets = [defaultConfig]) 
     uniformColorPalette,
     extendedSpacingScale,
     extendedFontSizeScale,
-    darkModeVariant,
     additionalBreakpoint,
   }
