+++ stubs/defaultConfig.stub.js
@@ -714,15 +714,15 @@ module.exports = {
     appearance: ['responsive'],
     backgroundAttachment: ['responsive'],
     backgroundClip: ['responsive'],
-    backgroundColor: ['responsive', 'hover', 'focus'],
+    backgroundColor: ['responsive', 'dark', 'hover', 'focus'],
     backgroundImage: ['responsive'],
-    gradientColorStops: ['responsive', 'hover', 'focus'],
+    gradientColorStops: ['responsive', 'dark', 'hover', 'focus'],
     backgroundOpacity: ['responsive', 'hover', 'focus'],
     backgroundPosition: ['responsive'],
     backgroundRepeat: ['responsive'],
     backgroundSize: ['responsive'],
     borderCollapse: ['responsive'],
-    borderColor: ['responsive', 'hover', 'focus'],
+    borderColor: ['responsive', 'dark', 'hover', 'focus'],
     borderOpacity: ['responsive', 'hover', 'focus'],
     borderRadius: ['responsive'],
     borderStyle: ['responsive'],
@@ -732,7 +732,7 @@ module.exports = {
     container: ['responsive'],
     cursor: ['responsive'],
     display: ['responsive'],
-    divideColor: ['responsive'],
+    divideColor: ['responsive', 'dark'],
     divideOpacity: ['responsive'],
     divideStyle: ['responsive'],
     divideWidth: ['responsive'],
@@ -775,7 +775,7 @@ module.exports = {
     placeContent: ['responsive'],
     placeItems: ['responsive'],
     placeSelf: ['responsive'],
-    placeholderColor: ['responsive', 'focus'],
+    placeholderColor: ['responsive', 'dark', 'focus'],
     placeholderOpacity: ['responsive', 'focus'],
     pointerEvents: ['responsive'],
     position: ['responsive'],
@@ -785,7 +785,7 @@ module.exports = {
     strokeWidth: ['responsive'],
     tableLayout: ['responsive'],
     textAlign: ['responsive'],
-    textColor: ['responsive', 'hover', 'focus'],
+    textColor: ['responsive', 'dark', 'hover', 'focus'],
     textOpacity: ['responsive', 'hover', 'focus'],
     textDecoration: ['responsive', 'hover', 'focus'],
     textTransform: ['responsive'],
